Restructure before main/test_main breakup
