+++ src/database/mongo/core.py
@@ -4,8 +4,8 @@ from pymongo.mongo_client import MongoClient
 from pymongo.server_api import ServerApi
 from bson import json_util
 
-from src.app.models.mongo.schemas import init_collections
 from src.config import MONGO_DATABASE_NAME
+from src.database.mongo.services import init_collections
 
 MONGO_URL = environ.get("CTI_MONGO_URL")
 if not MONGO_URL:
