+++ src/database/postgres/core.py
@@ -1,7 +1,6 @@
 from os import environ
 from dotenv import load_dotenv
 from sqlalchemy import create_engine
-from sqlalchemy.engine import URL
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
 
 # Note: We instantiate Base here because a single Base object will hold the Metadata
